Added update checker Moved config.rs and error.rs to new core dir Created mod.rs and updates.rs in core dir Changed imports to match new directory structure
