+++ src/core/mod.rs
@@ -1,0 +1,3 @@
+pub mod config;
+pub mod error;
+pub mod updates;
